#issue18 updated header component to support navigation
