+++ package.json
@@ -12,6 +12,7 @@
     "expo": "^30.0.1",
     "react": "16.3.1",
     "react-native": "https://github.com/expo/react-native/archive/sdk-30.0.0.tar.gz",
+    "react-navigation": "^2.17.0",
     "react-native-elements": "^0.19.1",
     "react-redux": "^5.0.7"
   }
